+++ precompiles/ics20/ICS20I.sol
@@ -2,6 +2,7 @@
 pragma solidity >=0.8.18;
 
 import "../common/Types.sol";
+import "../authorization/IICS20Authorization.sol";
 
 /// @dev The ICS20I contract's address.
 address constant ICS20_PRECOMPILE_ADDRESS = 0x0000000000000000000000000000000000000802;
@@ -19,19 +20,29 @@ struct DenomTrace {
     string baseDenom;
 }
 
-/// @dev Allocation represents a single allocation for an IBC fungible token transfer.
-struct Allocation {
-    string sourcePort;
-    string sourceChannel;
-    Coin[] spendLimit;
-    string[] allowList;
-}
-
 /// @author Evmos Team
 /// @title ICS20 Transfer Precompiled Contract
 /// @dev The interface through which solidity contracts will interact with IBC Transfer (ICS20)
 /// @custom:address 0x0000000000000000000000000000000000000802
-interface ICS20I {
+interface ICS20I is IICS20Authorization {
+    /// @dev Emitted when an ICS-20 transfer is executed.
+    /// @param sender The address of the sender.
+    /// @param receiver The address of the receiver.
+    /// @param sourcePort The source port of the IBC transaction.
+    /// @param sourceChannel The source channel of the IBC transaction.
+    /// @param denom The denomination of the tokens transferred.
+    /// @param amount The amount of tokens transferred.
+    /// @param memo The IBC transaction memo.
+    event IBCTransfer(
+        address indexed sender,
+        string indexed receiver,
+        string sourcePort,
+        string sourceChannel,
+        string denom,
+        uint256 amount,
+        string memo
+    );
+
     /// @dev Transfer defines a method for performing an IBC transfer.
     /// @param sourcePort the port on which the packet will be sent
     /// @param sourceChannel the channel by which the packet will be sent
@@ -77,111 +88,4 @@ interface ICS20I {
         string memory trace
     ) external view returns (string memory hash);
 
-    /// @dev Approves IBC transfer with a specific amount of tokens.
-    /// @param grantee The address for which the transfer authorization is granted.
-    /// @param allocations the allocations for the authorization.
-    function approve(
-        address grantee,
-        Allocation[] calldata allocations
-    ) external returns (bool approved);
-
-    /// @dev Revokes IBC transfer authorization for a specific grantee.
-    /// @param grantee The address for which the transfer authorization will be revoked.
-    function revoke(address grantee) external returns (bool revoked);
-
-    /// @dev Returns the remaining number of tokens that a grantee smart contract
-    /// will be allowed to spend on behalf of granter through
-    /// IBC transfers. This is an empty by array.
-    /// @param grantee The address of the contract that is allowed to spend the granter's tokens.
-    /// @param granter The address of the account able to transfer the tokens.
-    /// @return allocations The remaining amounts allowed to spend for
-    /// corresponding source port and channel.
-    function allowance(
-        address grantee,
-        address granter
-    ) external view returns (Allocation[] memory allocations);
-
-    /// @dev Increase the allowance of a given grantee by a specific amount of tokens for IBC transfer methods.
-    /// @param grantee The address of the contract that is allowed to spend the granter's tokens.
-    /// @param sourcePort the port on which the packet will be sent
-    /// @param sourceChannel the channel by which the packet will be sent
-    /// @param denom the denomination of the Coin to be transferred to the receiver
-    /// @param amount The amount of tokens to be spent.
-    /// @return approved is true if the operation ran successfully
-    function increaseAllowance(
-        address grantee,
-        string calldata sourcePort,
-        string calldata sourceChannel,
-        string calldata denom,
-        uint256 amount
-    ) external returns (bool approved);
-
-    /// @dev Decreases the allowance of a given grantee by a specific amount of tokens for for IBC transfer methods.
-    /// @param grantee The address of the contract that is allowed to spend the granter's tokens.
-    /// @param sourcePort the port on which the packet will be sent
-    /// @param sourceChannel the channel by which the packet will be sent
-    /// @param denom the denomination of the Coin to be transferred to the receiver
-    /// @param amount The amount of tokens to be spent.
-    /// @return approved is true if the operation ran successfully
-    function decreaseAllowance(
-        address grantee,
-        string calldata sourcePort,
-        string calldata sourceChannel,
-        string calldata denom,
-        uint256 amount
-    ) external returns (bool approved);
-
-    /// @dev Emitted when an ICS-20 transfer is executed.
-    /// @param sender The address of the sender.
-    /// @param receiver The address of the receiver.
-    /// @param sourcePort The source port of the IBC transaction.
-    /// @param sourceChannel The source channel of the IBC transaction.
-    /// @param denom The denomination of the tokens transferred.
-    /// @param amount The amount of tokens transferred.
-    /// @param memo The IBC transaction memo.
-    event IBCTransfer(
-        address indexed sender,
-        string indexed receiver,
-        string sourcePort,
-        string sourceChannel,
-        string denom,
-        uint256 amount,
-        string memo
-    );
-
-    /// @dev Emitted when an ICS-20 transfer authorization is granted.
-    /// @param grantee The address of the grantee.
-    /// @param granter The address of the granter.
-    /// @param sourcePort The source port of the IBC transaction.
-    /// @param sourceChannel The source channel of the IBC transaction.
-    /// @param spendLimit The coins approved in the allocation
-    event IBCTransferAuthorization(
-        address indexed grantee,
-        address indexed granter,
-        string sourcePort,
-        string sourceChannel,
-        Coin[] spendLimit
-    );
-
-    /// @dev This event is emitted when an granter revokes a grantee's allowance.
-    /// @param grantee The address of the grantee.
-    /// @param granter The address of the granter.
-    event RevokeIBCTransferAuthorization(
-        address indexed grantee,
-        address indexed granter
-    );
-
-    /// @dev This event is emitted when the allowance of a grantee is changed by a call to the decrease or increase
-    /// allowance method. The values field specifies the new allowances and the methods field holds the
-    /// information for which methods the approval was set.
-    /// @param grantee The address of the grantee.
-    /// @param granter The address of the granter.
-    /// @param methods The message type URLs of the methods for which the approval is set.
-    /// @param values The amounts of tokens approved to be spent.
-    event AllowanceChange(
-        address indexed grantee,
-        address indexed granter,
-        string[] methods,
-        uint256[] values
-    );
 }